fix(core): unify mfa factors sorting
